refactor app views into routes
