+++ app/app/settings/page.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import { SettingsView } from "../AppShell";
+import BottomNav from "@/components/BottomNav";
+
+export default function SettingsPage() {
+  return (
+    <>
+      <SettingsView />
+      <BottomNav value="settings" />
+    </>
+  );
+}
