+++ components/BottomNav.tsx
@@ -6,31 +6,23 @@ import { Leaf, Sprout, BarChart3, Cog, History } from "lucide-react";
 
 export type Tab = "today" | "timeline" | "plants" | "insights" | "settings";
 
-export default function BottomNav({
-  value,
-  onChange,
-}: {
-  value: Tab;
-  onChange?: (t: Tab) => void;
-}) {
+export default function BottomNav({ value }: { value: Tab }) {
   const paths: Record<Tab, string> = {
-    today: "/app?tab=today",
-    timeline: "/app?tab=timeline",
-    plants: "/app?tab=plants",
+    today: "/app/today",
+    timeline: "/app/timeline",
+    plants: "/app/plants",
     insights: "/app/insights",
-    settings: "/app?tab=settings",
+    settings: "/app/settings",
   };
 
   const Item = ({
     tab,
     label,
     icon,
-    href,
   }: {
     tab: Tab;
     label: string;
     icon: React.ReactNode;
-    href?: string;
   }) => {
     const active = value === tab;
     const className = `py-3 flex flex-col items-center justify-center text-xs ${
@@ -38,30 +30,14 @@ export default function BottomNav({
         ? "text-neutral-900 dark:text-neutral-100"
         : "text-neutral-600 dark:text-neutral-400"
     }`;
-
-    if (href) {
-      return (
-        <Link href={href} className={className} aria-current={active ? "page" : undefined}>
-          <div className={`mb-1 ${active ? "" : "opacity-80"}`}>{icon}</div>
-          <span className="font-sans">{label}</span>
-          {active && (
-            <span className="mt-1 h-1 w-6 rounded-full bg-neutral-900 dark:bg-neutral-100" />
-          )}
-        </Link>
-      );
-    }
     return (
-      <button
-        onClick={() => onChange?.(tab)}
-        className={className}
-        aria-current={active ? "page" : undefined}
-      >
+      <Link href={paths[tab]} className={className} aria-current={active ? "page" : undefined}>
         <div className={`mb-1 ${active ? "" : "opacity-80"}`}>{icon}</div>
         <span className="font-sans">{label}</span>
         {active && (
           <span className="mt-1 h-1 w-6 rounded-full bg-neutral-900 dark:bg-neutral-100" />
         )}
-      </button>
+      </Link>
     );
   };
 
@@ -71,11 +47,11 @@ export default function BottomNav({
       style={{ paddingBottom: "env(safe-area-inset-bottom)" }}
     >
       <div className="max-w-screen-sm mx-auto grid grid-cols-5">
-        <Item tab="today" label="Today" icon={<Leaf />} href={!onChange ? paths.today : undefined} />
-        <Item tab="timeline" label="Timeline" icon={<History />} href={!onChange ? paths.timeline : undefined} />
-        <Item tab="plants" label="Plants" icon={<Sprout />} href={!onChange ? paths.plants : undefined} />
-        <Item tab="insights" label="Insights" icon={<BarChart3 />} href={paths.insights} />
-        <Item tab="settings" label="Settings" icon={<Cog />} href={!onChange ? paths.settings : undefined} />
+        <Item tab="today" label="Today" icon={<Leaf />} />
+        <Item tab="timeline" label="Timeline" icon={<History />} />
+        <Item tab="plants" label="Plants" icon={<Sprout />} />
+        <Item tab="insights" label="Insights" icon={<BarChart3 />} />
+        <Item tab="settings" label="Settings" icon={<Cog />} />
       </div>
     </nav>
   );
